feat(dashboard): verify Home route integration
Consolidates Home onto the active page route and covers the landing page,
activity feed, workspace cards, and registry sections in frontend tests.
